services/v1/notification/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
requests==2.23.0
Flask-Limiter==1.4
boto3==1.14.59
mailjet-rest==1.3.4